requirements.txt
PyMySQL==1.1.0
future==0.18.2
configparser==5.0.2
Flask==2.1.0
line-bot-sdk==1.20.0
pandas==1.3.3
gunicorn
openai
python-dotenv